+++ plugin/includes/class-m360-core.php
@@ -21,6 +21,7 @@ require_once M360_CORE_PATH . 'includes/ads/class-m360-ads-admin.php';
 require_once M360_CORE_PATH . 'includes/ads/class-m360-ads-creatives-admin.php';
 require_once M360_CORE_PATH . 'includes/search/class-m360-search-controller.php';
 require_once M360_CORE_PATH . 'includes/search/class-m360-search-form-component.php';
+require_once M360_CORE_PATH . 'includes/header/class-m360-header-orchestrator.php';
 require_once M360_CORE_PATH . 'includes/author/class-m360-author-controller.php';
 require_once M360_CORE_PATH . 'includes/category/class-m360-category-controller.php';
 require_once M360_CORE_PATH . 'includes/tag/class-m360-tag-controller.php';
@@ -58,6 +59,7 @@ final class M360_Core_Runtime_034
         M360_Ads_DB::maybe_upgrade();
         M360_Ads_Admin::register();
         M360_Ads_Creatives_Admin::register();
+        M360_Header_Orchestrator::register();
         M360_Ads_Inline_Engine::register();
         $this->init_view_engine();
         add_action('init', [$this, 'register_shortcodes']);
@@ -99,6 +101,7 @@ final class M360_Core_Runtime_034
         wp_register_style('m360-core-search', M360_CORE_URL . 'assets/css/search.css', ['m360-core-ui-polish', 'm360-core-ui-components', 'm360-core-navigation-components'], M360_CORE_VERSION);
         wp_register_style('m360-core-search-form', M360_CORE_URL . 'assets/css/m360-search-form.css', ['m360-core-foundation'], M360_CORE_VERSION);
         wp_register_script('m360-core-search-form', M360_CORE_URL . 'assets/js/m360-search-form.js', [], M360_CORE_VERSION, true);
+        wp_register_style('m360-core-header-orchestrator', M360_CORE_URL . 'assets/css/m360-header-orchestrator.css', ['m360-core-ads', 'm360-core-search-form'], M360_CORE_VERSION);
         wp_register_style('m360-core-author', M360_CORE_URL . 'assets/css/author.css', ['m360-core-ui-polish', 'm360-core-ui-components', 'm360-core-navigation-components'], M360_CORE_VERSION);
         wp_register_style('m360-core-category', M360_CORE_URL . 'assets/css/category.css', ['m360-core-ui-polish', 'm360-core-ui-components', 'm360-core-navigation-components'], M360_CORE_VERSION);
         wp_register_style('m360-core-tag', M360_CORE_URL . 'assets/css/tag.css', ['m360-core-ui-polish', 'm360-core-ui-components', 'm360-core-navigation-components'], M360_CORE_VERSION);
@@ -121,6 +124,7 @@ final class M360_Core_Runtime_034
         M360_Navigation_Shortcodes::register();
         M360_Post_Info_Component::register_shortcodes();
         M360_Search_Form_Component::register_shortcodes();
+        M360_Header_Orchestrator::register_shortcodes();
         M360_UI_Components::register_shortcodes();
         M360_Latest_News_Component::register_shortcodes();
         M360_Slot_Renderer::register_shortcodes();